+++ packages/cmrjs/tests/cmr-utils/test-cmr-utils.js
@@ -106,65 +106,44 @@ test.after.always(async (t) => {
   t.context.launchpadStub.restore();
 });
 
-test('isCMRFile returns truthy if fileobject has valid xml name', (t) => {
-  const fileObj = {
-    name: 'validfile.cmr.xml',
-  };
-  t.truthy(isCMRFile(fileObj));
-});
-
-test('isCMRFile returns falsy if fileobject does not valid xml name', (t) => {
-  const fileObj = {
-    name: 'invalidfile.xml',
-  };
-  t.falsy(isCMRFile(fileObj));
-});
-
-test('isCMRFile returns truthy if fileobject has valid json name', (t) => {
-  const fileObj = {
-    name: 'validfile.cmr.json',
-  };
-  t.truthy(isCMRFile(fileObj));
-});
-
-test('isCMRFile returns falsy if fileobject does not valid json name', (t) => {
-  const fileObj = {
-    name: 'invalidfile.json',
-  };
-  t.falsy(isCMRFile(fileObj));
-});
-
-test('isCMRFile returns truthy if fileobject has valid xml filenamename', (t) => {
+test('isCMRFile returns true if fileobject has valid .cmr.xml filename', (t) => {
   const fileObj = {
     filename: 'validfile.cmr.xml',
   };
-  t.truthy(isCMRFile(fileObj));
+  t.true(isCMRFile(fileObj));
 });
 
-test('isCMRFile returns falsy if fileobject does not valid xml filenamename', (t) => {
+test('isCMRFile returns false if fileobject does not have a valid xml filename', (t) => {
   const fileObj = {
     filename: 'invalidfile.xml',
   };
-  t.falsy(isCMRFile(fileObj));
+  t.false(isCMRFile(fileObj));
 });
 
-test('isCMRFile returns truthy if fileobject has valid json filenamename', (t) => {
+test('isCMRFile returns true if fileobject has a valid cmr_iso.xml filename', (t) => {
+  const fileObj = {
+    filename: 'validfile.cmr_iso.xml',
+  };
+  t.true(isCMRFile(fileObj));
+});
+
+test('isCMRFile returns true if fileobject has valid json filename', (t) => {
   const fileObj = {
     filename: 'validfile.cmr.json',
   };
-  t.truthy(isCMRFile(fileObj));
+  t.true(isCMRFile(fileObj));
 });
 
-test('isCMRFile returns falsy if fileobject does not valid json filenamename', (t) => {
+test('isCMRFile returns false if fileobject does not valid json filename', (t) => {
   const fileObj = {
     filename: 'invalidfile.json',
   };
-  t.falsy(isCMRFile(fileObj));
+  t.false(isCMRFile(fileObj));
 });
 
-test('isCMRFile returns falsy if fileobject is invalid', (t) => {
+test('isCMRFile returns false if fileobject is invalid', (t) => {
   const fileObj = { bad: 'object' };
-  t.falsy(isCMRFile(fileObj));
+  t.false(isCMRFile(fileObj));
 });
 
 test('constructCmrConceptLink returns echo10 link', (t) => {
@@ -319,7 +298,7 @@ test.serial('updateEcho10XMLMetadata adds granule files correctly to OnlineAcces
       metadataObject.Granule.AssociatedBrowseImageUrls.ProviderBrowseUrl,
       AssociatedBrowseExpected
     );
-    t.truthy(uploadEchoSpy.calledWith('testXmlString',
+    t.true(uploadEchoSpy.calledWith('testXmlString',
       { filename: 's3://cumulus-test-sandbox-private/notUsed' }));
   } finally {
     revertMetaObject();
@@ -398,7 +377,6 @@ test.serial('updateUMMGMetadata adds Type correctly to RelatedURLs for granule f
 test.serial('getGranuleTemporalInfo returns temporal information from granule CMR json file', async (t) => {
   const cmrJSON = await fs.readFile('./tests/fixtures/MOD09GQ.A3411593.1itJ_e.006.9747594822314.cmr.json', 'utf8');
   const cmrMetadata = JSON.parse(cmrJSON);
-  const revertCmrFileObject = cmrUtil.__set__('granuleToCmrFileObject', () => ([{ filename: 'test.cmr.json', granuleId: 'testGranuleId' }]));
   const revertMetaObject = cmrUtil.__set__('metadataObjectFromCMRJSONFile', () => cmrMetadata);
 
   const expectedTemporalInfo = {
@@ -411,12 +389,13 @@ test.serial('getGranuleTemporalInfo returns temporal information from granule CM
   try {
     const temporalInfo = await getGranuleTemporalInfo({
       granuleId: 'testGranuleId',
-      files: [],
+      files: [{
+        filename: 'test.cmr.json',
+      }],
     });
 
     t.deepEqual(temporalInfo, expectedTemporalInfo);
   } finally {
-    revertCmrFileObject();
     revertMetaObject();
   }
 });
@@ -424,7 +403,6 @@ test.serial('getGranuleTemporalInfo returns temporal information from granule CM
 test.serial('getGranuleTemporalInfo returns temporal information from granule CMR xml file', async (t) => {
   const cmrXml = await fs.readFile('./tests/fixtures/cmrFileUpdateFixture.cmr.xml', 'utf8');
   const cmrMetadata = await (promisify(xml2js.parseString))(cmrXml, xmlParseOptions);
-  const revertCmrFileObject = cmrUtil.__set__('granuleToCmrFileObject', () => ([{ filename: 'test.cmr.xml', granuleId: 'testGranuleId' }]));
   const revertMetaObject = cmrUtil.__set__('metadataObjectFromCMRXMLFile', () => cmrMetadata);
 
   const expectedTemporalInfo = {
@@ -437,12 +415,39 @@ test.serial('getGranuleTemporalInfo returns temporal information from granule CM
   try {
     const temporalInfo = await getGranuleTemporalInfo({
       granuleId: 'testGranuleId',
-      files: [],
+      files: [{
+        filename: 'test.cmr.xml',
+      }],
     });
 
     t.deepEqual(temporalInfo, expectedTemporalInfo);
   } finally {
-    revertCmrFileObject();
+    revertMetaObject();
+  }
+});
+
+test.serial('getGranuleTemporalInfo returns temporal information from granule CMR ISO xml file', async (t) => {
+  const cmrXml = await fs.readFile('./tests/fixtures/ATL03_20190101000418_00540208_004_01.cmr_iso.xml', 'utf8');
+  const cmrMetadata = await (promisify(xml2js.parseString))(cmrXml, xmlParseOptions);
+  const revertMetaObject = cmrUtil.__set__('metadataObjectFromCMRXMLFile', () => cmrMetadata);
+
+  const expectedTemporalInfo = {
+    beginningDateTime: '2019-01-01T00:04:18.809303Z',
+    endingDateTime: '2019-01-01T00:11:20.899913Z',
+    productionDateTime: '2021-02-05T04:23:58.000000Z',
+    lastUpdateDateTime: '2021-05-07T09:10:59.891292Z',
+  };
+
+  try {
+    const temporalInfo = await getGranuleTemporalInfo({
+      granuleId: 'testGranuleId',
+      files: [{
+        filename: 'test.cmr_iso.xml',
+      }],
+    });
+
+    t.deepEqual(temporalInfo, expectedTemporalInfo);
+  } finally {
     revertMetaObject();
   }
 });